split telemetry event classes into own files
