+++ lib/datadog/core/telemetry/event/app_closing.rb
@@ -1,0 +1,14 @@
+module Datadog
+  module Core
+    module Telemetry
+      class Event
+        # Telemetry class for the 'app-closing' event
+        class AppClosing < Base
+          def type
+            'app-closing'
+          end
+        end
+      end
+    end
+  end
+end
