+++ lib/datadog/core/telemetry/event/app_dependencies_loaded.rb
@@ -1,0 +1,30 @@
+module Datadog
+  module Core
+    module Telemetry
+      class Event
+        # Telemetry class for the 'app-dependencies-loaded' event
+        class AppDependenciesLoaded < Base
+          def type
+            'app-dependencies-loaded'
+          end
+
+          def payload
+            { dependencies: dependencies }
+          end
+
+          private
+
+          def dependencies
+            Gem.loaded_specs.collect do |name, gem|
+              {
+                name: name,
+                version: gem.version.to_s,
+                # hash: nil,
+              }
+            end
+          end
+        end
+      end
+    end
+  end
+end
